Daraja API integration (WIP)

+++ src/main/java/com/example/Pharmacy/Application/order/model/OrderItem.java
@@ -8,6 +8,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -21,7 +22,7 @@ public class OrderItem {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long orderItemId;
     private @NotNull int quantity;
-    private @NotNull double price;
+    private @NotNull BigDecimal price;
     private Date createdDate;
 
     @ManyToOne
